appview/pyproject.toml
[project]
name = "thedistance-appview"
version = "0.1.0"
description = "AppView backend for The Distance"
requires-python = ">=3.12"
dependencies = [
  "websockets>=15.0",
  "fastapi>=0.115",
  "uvicorn>=0.34",
  "psycopg[binary]>=3.2",
  "python-dotenv>=1.0",
  "pydantic-settings>=2.6",
  "httpx>=0.28",
]

[project.scripts]
start = "app.cli:start"
subscribe = "app.cli:subscribe"
backfill = "app.cli:backfill"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
package = true

[dependency-groups]
dev = ["ruff>=0.8"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
